Social Media Log in Clone
This is just a quick demo to show how easy it is to “clone” a popular
trusted site for the purpose of tricking users into putting in their
information.

Don’t do anything dumb with this info. I won’t be held responsible if
you use this for anything illegal.